remove ycm, unnecessary omnifunc mappings
Learned that you need to cd into the current directory before running
the vim-go run command in order for it to work correctly. Use <leader>cd
to change cwd to same as current buffer location.

YCM is uncessary since release of gopls. Switched godef and goinfo to
use gopls. Unfortunately, everything works with gopls except refactoring
capabilities, so files still need to be in GOPATH in order to use
gorename for refactoring.