changed initialized var and path mutation
Switched path mutation from appending to path to adding to
`fish_user_paths`. This allows the reload function to work properly
since the fish_user_paths will not continue to add to the PATH if it is
already present. It will just move it to the front of the path.