Explored python coc settings for conda
Figured out that in order for coc to use conda's environments you need
to specify the created conda env as the "python.pythonPath". So first
create the conda env via conda create and then update the coc config
prior to working on a particular python project. For some reason the
builtin coc command for selecting a particular interpreter is not
working as expected for conda envs.