Removed autosave plugin, updated go mappings
Realized that the autosave plugin was a bit too aggressive and has some
performance degredation without much added utility, especially since I
originally got it for the coc rename refactoring for classes not saving
the buffers after editing. Which is also not the best so I just use the
Rename command from eunuch and the grepper for global refactoring.