Added function to clean eclipse workspace for windows
The coc.nvim plugin running on windows was having issues when trying to
clean the jdt.ls workspace. It would automatically bypass the prompt
asking if the user was sure they wanted to clean the workspace. This
function can be used to remove all of the files and directories created
by the jdt.ls server, so that the workspace can be regenerated.