Removed conda from fish cfg, fixed win vimfiles var and airline refresh
Removed conda init from fish config file to reduce shell startup time.
Also I am not currently using anaconda so its not necessary to have the
conda command accessible in the PATH.

Windows _vimfiles directory changed to vimfiles.

Refreshing the vimrc/init.vim files to reload the vim config caused the
airline plugin's tabline to disappear. This was due to the stal setting
in the vimrc. Removing this setting allows the removal of the toggle
hack.