Removing gutentags_plus, jdk-src in project
It appears that when dealing with multiple gtags-cscope databases, the
first takes precendence and the other files are referenced with
relatives paths to the first, which throws off any chance of navigating
to them. This occurs in both cache dir and project dir. So I disabled
the auto connect flag. Also, gutentags_plus was changing a lot of
settings under the hood which I dont appreciate.

In order to get the jdk src files in the database, the tags need to be
computed with the jdk files in the project dir. Since I dont want them
in vcs, I had to changed the gutentag_file_list_command = 'fd -t f -e java'
to get all java files regardless of if they are ignored by the vcs.

I tried using the $GTAGSLIBPATH and I think it worked but each project
using those jdk files would also have to be recomputed and it wasnt
working switching the env var inside vim.

In the future I would like to try and change the list command to gather
from multiple sources and aggregate the list in order to keep the jdk /
std lib files separate from the project files.